sync_python/pyproject.toml
[tool.poetry]
name = "sync_python"
version = "0.1.0"
description = ""
authors = ["Steve Dignam <steve@dignam.xyz>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
gunicorn = "^20.1.0"
dj-database-url = "^1.0.0"
psycopg2 = "^2.9.3"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"
orjson = "^3.8.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
